Use hexpm/elixir base image
The `hexpm/elixir` images have a better tagging system and are provided
by the Hex team itself.

These images do not contain elixir & OTP sources, so navigating to
standard library source files will not be possible. This is a reasonable
trade-off, as it was actually not possible in the previous builds either
- not because the source wasn't available in the container, but because
the editor can't open files inside the container. The previous UX was
actually worse, as it would just open an empty file when you tried to
go to the definition of a function in the standard library. Now it tells
you that it's not possible to do so.

I dropped the checksum check, because the code provided in the download
is pretty much just a proxy to download the actual elixir-ls code. I
consider the checksum check to have little value in this context.

I tried to build from source as well, which worked, but every time I
launched elixir-ls, it would recompile `language-server` module. This
added unnecessary time to launch. I believe it's because `Mix.install/2`
is used, which triggers a recompile. It would be possible to patch the
code in elixir-ls to not do this, but it didn't seem worth it to hack
that together.